Moved the function to services to improve modularity
